Refactor to model ListBoxAutomationPeer after ComboBoxAutomationPeer
These serve similar behaviors, and the unrealized/virtual element peer can
subclass from UnrealizedElementAutomationPeer.
